wm: Fade screen on signout.
Upon receiving a message from Chrome letting us know that
the user is signing out, we grab the pointer (to hide the
cursor) and make the screen fade quickly to black.

BUG=chromium-os:9915
TEST=added; also built an image and checked that the previous shutting-down animation still works pre- and post-login and that the new animation is shown when signing out

Review URL: http://codereview.chromium.org/6011011

+++ real_x_connection.cc
@@ -315,7 +315,7 @@ bool RealXConnection::GrabPointer(XWindow xid,
 
   if (reply->status != XCB_GRAB_STATUS_SUCCESS) {
     LOG(WARNING) << "Pointer grab for window " << XidStr(xid)
-                 << " returned status " << reply->status;
+                 << " returned status " << static_cast<int>(reply->status);
     return false;
   }
   return true;
@@ -348,7 +348,7 @@ bool RealXConnection::GrabKeyboard(XWindow xid, XTime timestamp) {
 
   if (reply->status != XCB_GRAB_STATUS_SUCCESS) {
     LOG(WARNING) << "Keyboard grab for window " << XidStr(xid)
-                 << " returned status " << reply->status;
+                 << " returned status " << static_cast<int>(reply->status);
     return false;
   }
   return true;
